Testing for project and its features done
